Added Login Page and Function
I added a login page used from my CSE108 final project and got some code for the logic. The session is for a join room function, which can be used for the current session to make a current session to keep track of what they add to the lists alongside other to be saved for end users to check later.